Added EC2 instance and EBS details to README
Expanded the README documentation with additional sections on what is needed for an Amazon EC2 instance and details about Amazon EBS (Elastic Block Store). The new content includes a list of requirements for setting up an EC2 instance, as well as a description of EBS and its various volume types.